+++ analytics/src/main/java/com/rakuten/tech/mobile/analytics/Tracker.kt
@@ -2,7 +2,9 @@ package com.rakuten.tech.mobile.analytics
 
 /** Event tracking dummy class. Please notice the proguard class path. */
 class Event(name: String, parameters: Map<String, Any>) {
-    fun track() {}
+    fun track() {
+        //do nothing intended
+    }
 }
 
 object RatTracker {