+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/SettingsNavigator.kt
@@ -1,5 +1,0 @@
-package com.rakuten.tech.mobile.testapp.ui.settings
-
-interface SettingsNavigator {
-    fun navigateToScreen(): Boolean
-}