+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **Feature:** Added `sendJsonToMiniApp` in `MiniAppView` and `MiniAppViewImpl` for receiving messages from the HostApp to MiniApps.
 - **Fix:** Prevent file creation with zero byte while downloading a file using `MiniAppFileDownloader` in Android 10.
 - **Fix:** Prevent file creation with zero byte while downloading a file using `MiniAppFileDownloader` in the offline mode.
+- **Fix:** Prevent file creation with zero byte with appending a single dot(.) at the end of file's name while downloading a file using `MiniAppFileDownloader` in Android 10.
 
 **Sample App**
 - **Feature:** Settings is moved to it's own tab.